+++ now/apps/text_to_text/ft-flow-sbert.yml
@@ -52,7 +52,7 @@ executors:
     env:
       JINA_LOG_LEVEL: DEBUG
   - name: postprocessor
-    uses: jinahub+docker://NOWPostprocessorV2/v0.0.6
+    uses: jinahub+docker://NOWPostprocessor/v0.0.5
     env:
       JINA_LOG_LEVEL: DEBUG
   - name: indexer
