+++ now/apps/text_to_video/ft-flow-video-clip-high-performance.yml
@@ -58,7 +58,7 @@ executors:
     env:
       JINA_LOG_LEVEL: DEBUG
   - name: postprocessor
-    uses: jinahub+docker://NOWPostprocessorV2/v0.0.6
+    uses: jinahub+docker://NOWPostprocessor/v0.0.5
     uses_with:
       traversal_paths: '@c'
     env:
