+++ now/constants.py
@@ -5,8 +5,8 @@ from now.utils import BetterEnum
 # DEMO_DATASET_DOCARRAY_VERSION = docarray_version
 DEMO_DATASET_DOCARRAY_VERSION = '0.13.17'
 
-DOCKER_BFF_PLAYGROUND_TAG = '0.0.122-auth-exec-4'
-NOW_PREPROCESSOR_VERSION = '0.0.82-test-preprocessing'
+DOCKER_BFF_PLAYGROUND_TAG = '0.0.123-filtering-playground-1'
+NOW_PREPROCESSOR_VERSION = '0.0.83-filtering-playground-1'
 NOW_AUTH_EXECUTOR_VERSION = '0.0.5-auth-exec-4'
 NOW_ANNLITE_INDEXER_VERSION = '0.0.3-annlite-secure-1'
