+++ now_executors/NOWPostprocessor/config.yml
@@ -1,4 +1,4 @@
-jtype: NOWPostprocessorV2
+jtype: NOWPostprocessor
 metas:
   py_modules:
     - executor.py
