+++ now_executors/NOWPostprocessor/manifest.yml
@@ -1,3 +1,3 @@
 manifest_version: 1
-name: NOWPostprocessorV2
+name: NOWPostprocessor
 description: Post-processes any documents after encoding such that they are ready to be indexed, used as query, etc.
